+++ Changelog.md
@@ -6,6 +6,8 @@ Released on ?.
 
 ✨ New Features / Improvements
 
+- Added long-tap map menu to find flights, opening xcontest worldwide flights
+  search
 - Android: `geo:` links can now be opened with the app and adds a new location
 - Improved recognizing more geo coordinate formats
 - Minimum Android version is now Android 12